Can't have datetime values with a time but no year.
(The second datetime value is actually a bit problematic. It's clear
from context that "30th of October. Midnight." actually means midnight
following the 30th, so technically the datetime value should be
"10-31". But having that wrapped around the string "30th of October"
is just a recipe for confusion. Frankly, I think the proper solution
would be to remove the time tags altogether here, but that doesn't
seem to be an option.)